backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.2.1+cpu
ultralytics==8.3.128
pillow==10.1.0
python-dotenv==1.0.0
numpy==1.24.3 